Move to a "slot based" profiler
Where we can now also accumulate profile results to a "slot", meaning
that, for example, the ProfParseFloat slot will contain the total time
spent in ParseFloat.

But it is rather unsatisfactory as the comment in the code points out:
there's a significant overhead to these calls in small functions.

And we have not solved the problem of calculating the non-overlapping
times of functions, or of handling recursive calls.