Profile recursive blocks
Modify the profiler so that it supports recursive blocks. This was done
by giving each profiling block its own data structure, and only writing
back to the global result at the end of each block.

Move cputicks functions to separate package.

Remove pprof, the whole point is that we are writing our own profiler.