Implement the test probing for cache sizes
Implement the test that reads ever increasing amounts of data from the
input buffer, to see where the drop offs in bandwidth occurs.

On my laptop with an i7-8550U I am seeing larger drops between:

  * 32 KB and 64 KB,
  * 256 KB and 512 KB,
  * 4 MB and 8 MB and,
  * 8 MB and 16 MB.

The drop off between 4 MB and 8 MB is there but is not as pronounced as
the others. Not sure what that is about. According to `lstopo` my CPU
has 32 KB L1i cache, 256 KB L2 cache and 8 MB L3 cache, which explains
the numbers above.

Also not sure if I need to unroll the vmovdqu instructions this much,
but decreasing the number also makes the bandwidth go down quite a bit.
Inserting a `PCALIGN $64` in the Go assembly files brings it somewhat
back up, but that cannot currently be generated with avo, making it a
bit of a hassle.